refactor: centralize mock data
