--- frontend-next/src/data/mockData.json
@@ -1,34 +1,0 @@
-{
-  "metrics": {
-    "steps": 12345,
-    "resting_hr": 60,
-    "vo2max": 50,
-    "sleep_hours": 7.5
-  },
-  "stepsHistory": [10000, 11000, 9000, 12000, 8000, 13000, 10000],
-  "hrZones": [20, 40, 30, 10],
-  "sleepStages": [3, 4, 2, 1],
-  "vo2History": [49, 50, 50, 50, 51, 50, 50],
-  "activities": [
-    { "time": "2024-05-01T00:00:00Z", "steps": 10000, "resting_hr": 60, "vo2max": 50, "sleep_hours": 7.0 },
-    { "time": "2024-05-02T00:00:00Z", "steps": 11000, "resting_hr": 61, "vo2max": 50, "sleep_hours": 7.2 },
-    { "time": "2024-05-03T00:00:00Z", "steps": 9000, "resting_hr": 59, "vo2max": 49, "sleep_hours": 6.8 },
-    { "time": "2024-05-04T00:00:00Z", "steps": 12000, "resting_hr": 60, "vo2max": 50, "sleep_hours": 7.4 },
-    { "time": "2024-05-05T00:00:00Z", "steps": 8000, "resting_hr": 58, "vo2max": 51, "sleep_hours": 8.0 },
-    { "time": "2024-05-06T00:00:00Z", "steps": 13000, "resting_hr": 57, "vo2max": 50, "sleep_hours": 7.8 },
-    { "time": "2024-05-07T00:00:00Z", "steps": 10000, "resting_hr": 59, "vo2max": 50, "sleep_hours": 7.6 }
-  ],
-  "goals": {
-    "steps": 10000,
-    "sleep_hours": 8
-  },
-  "gps": {
-    "coordinates": [
-      [37.7749, -122.4194],
-      [37.775, -122.418],
-      [37.776, -122.417],
-      [37.777, -122.4185],
-      [37.7765, -122.42]
-    ]
-  }
-}
